+++ src/main/java/nosql/workshop/model/Installation.java
@@ -1,6 +1,7 @@
 package nosql.workshop.model;
 
 
+import com.google.gson.annotations.SerializedName;
 import org.jongo.marshall.jackson.oid.MongoId;
 
 import java.util.Date;
@@ -12,6 +13,7 @@ import java.util.List;
 public class Installation {
 
     @MongoId
+    @SerializedName("id")
     private String _id;
     private String nom;
     private Adresse adresse;
