+++ src/main/java/nosql/workshop/services/InstallationService.java
@@ -98,44 +98,4 @@ public class InstallationService {
         return stats;
     }
 
-    public List<Installation> search(String query) throws IOException {
-        List<Installation> installations = new ArrayList<Installation>();
-
-        JestClient client = ESConnectionUtil.createClient("");
-
-
-        String querySearch2 = "{\"query\": {\n"
-                + "     \"match\": {\n"
-                + "         \"nom\": \"" + query + "\"\n"
-                + "     }\n"
-                + "}}";
-
-
-        String querySearch3 = "{"
-            + " \"query\": {\n"
-            + "     \"multi_match\": {\n"
-            + "         \"query\": \"" + query + "\",\n"
-            + "         \"fields\": [\n"
-            + "             \"_all\"\n"
-            + "         ]\n"
-            + "     }\n"
-            + " }\n"
-            + "}";
-
-        String querySearch = "{\"query\": {\"multi_match\": {\"query\": \""+query+"\", \"fields\": [\"_all\"] }}}";
-        Search.Builder searchBuilder = new Search.Builder(querySearch).addIndex("installations").addType("installation");
-        System.out.println(querySearch);
-
-        SearchResult result = client.execute(searchBuilder.build());
-
-        List<SearchResult.Hit<Installation, Void>> hits = result.getHits(Installation.class);
-        Stream<Installation> installationStream = hits.stream().map(installation -> installation.source);
-        return Lists.newArrayList(installationStream.iterator());
-       /** if (!hits.isEmpty()) {
-            for (SearchResult.Hit<Installation, Void> hit : hits) {
-                installations.add(hit.source);
-            }
-        }
-        return installations;**/
-    }
 }
